* Fixes from the WestGrid people's use of Lintel/DataSeries
TODO: note that we have a conflict between use of off64_t and
  ExtentType::int64
Extent.H: Allow __x86_64__ to be defined and assume that it gives the
  appropriate fast byteswap routine -- this needs to be verified.
DataSeriesFile.C: Fix types to eliminate warning.
dsextentindex.C: Fix types to eliminate warning.
lvpvprocessusage.C: Fix types to hopefully eliminate warning.

+++ src/process/lvpvprocessusage.C
@@ -648,7 +648,7 @@ public:
 class rollupptr_hash {
 public:
     int operator()(const rollupptr &a) {
-	return a.id ^ (int)a.strval;
+	return a.id ^ (long)a.strval;
     }
 };
